+++ pandas/_libs/byteswap.pyi
