+++ setup.py
@@ -225,8 +225,8 @@ class CheckSDist(sdist_class):
         "pandas/_libs/tslibs/vectorized.pyx",
         "pandas/_libs/window/indexers.pyx",
         "pandas/_libs/writers.pyx",
-        "pandas/io/sas/sas.pyx",
-        "pandas/io/sas/byteswap.pyx",
+        "pandas/_libs/sas.pyx",
+        "pandas/_libs/byteswap.pyx",
     ]
 
     _cpp_pyxfiles = [
@@ -558,8 +558,8 @@ ext_data = {
     },
     "_libs.window.indexers": {"pyxfile": "_libs/window/indexers"},
     "_libs.writers": {"pyxfile": "_libs/writers"},
-    "io.sas._sas": {"pyxfile": "io/sas/sas"},
-    "io.sas._byteswap": {"pyxfile": "io/sas/byteswap"},
+    "_libs.sas": {"pyxfile": "_libs/sas"},
+    "_libs.byteswap": {"pyxfile": "_libs/byteswap"},
 }
 
 extensions = []
